requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
coverage==7.6.10
Django==4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
httplib2==0.22.0
idna==3.10
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-decouple==3.8
requests==2.32.3
rsa==4.9
ruff==0.9.2
sqlparse==0.5.3
taskipy==1.14.1
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
